Separate statistics logic in it's own file
